creating SDLogger class

+++ src/AnalogSensor.cpp
@@ -7,7 +7,7 @@ AnalogSensor::AnalogSensor(const uint8_t&pin, const bool &bidirectional, const f
 {
 	// Initialize current readings buffer
 	_readings = new float[_numReadings];
-	for (int i = 0; i < _numReadings; ++i)
+	for (uint8_t i = 0; i < _numReadings; ++i)
 	{
 		_readings[i] = 0;
 	}
@@ -34,7 +34,7 @@ void AnalogSensor::Calibrate()
 	uint8_t numCalibrationReadings = 100;
 	uint16_t calibrationTime = 500;
 
-	for (int i = 0; i < numCalibrationReadings; ++i)
+	for (uint8_t i = 0; i < numCalibrationReadings; ++i)
 	{
 		buffer += AnalogToValue();
 		delay(calibrationTime / numCalibrationReadings);
@@ -54,7 +54,7 @@ void AnalogSensor::Update()
 
 	// Calculate smoothed value
 	float buffer = 0;
-	for (int i = 0; i < _numReadings; ++i)
+	for (uint8_t i = 0; i < _numReadings; ++i)
 	{
 		buffer += _readings[i];
 	}
